requirements.txt
 # Web framework
Flask==2.3.2

# Database
SQLAlchemy==2.0.15

# API
requests==2.31.0

# Data processing
pandas==2.0.2
numpy==1.24.3

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.3.1

# Linting and formatting
flake8==6.0.0
black==23.3.0

# Security
bcrypt==4.0.1

# Async support
aiohttp==3.8.4

# Task queue
celery==5.3.0

# Caching
redis==4.5.5

# Logging
loguru==0.7.0

opencv-python
ultralytics
requests
python-dotenv
openai